#4567 rename package quarantine
https://github.com/metasfresh/metasfresh/issues/4567
